# restapi/Hotel.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Boutique Hotel Technikum – Booking API",
    "description": "REST API für die Buchungs-App des Hotel Technikum. Gäste können Hotelzimmer durchsuchen, Verfügbarkeit prüfen und Zimmer buchen.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/bookings": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Buchung erstellen",
        "description": "Erstellt eine neue Buchung für ein verfügbares Zimmer.\n\n**Validierungsregeln:**\n- `firstName`: nicht leer, max. 100 Zeichen\n- `lastName`: nicht leer, max. 100 Zeichen\n- `email`: gültiges E-Mail-Format\n- `roomId`: muss existierendes Zimmer referenzieren\n- `checkIn` darf nicht in der Vergangenheit liegen\n- `checkOut` muss nach `checkIn` liegen, oder gleich `checkIn` sein (gleicher Tag = 1 Nacht)\n- Maximaler Buchungszeitraum: 30 Nächte\n- Zimmer muss im gewählten Zeitraum verfügbar sein",
        "operationId": "createBooking",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              },
              "example": {
                "roomId": 1,
                "checkIn": "2026-07-01",
                "checkOut": "2026-07-05",
                "firstName": "Max",
                "lastName": "Mustermann",
                "email": "max@example.com",
                "breakfast": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Buchung erfolgreich erstellt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                },
                "example": {
                  "id": "a3f1c2d4-e5b6-7890-abcd-ef1234567890",
                  "room": {
                    "id": 1,
                    "title": "Deluxe Suite",
                    "description": "Geräumige Suite mit Balkon und Bergblick.",
                    "imageUrl": "https://hotel-technikum.at/images/rooms/1.jpg",
                    "pricePerNight": 149.0,
                    "extras": [
                      {
                        "id": 1,
                        "name": "WiFi",
                        "icon": "wifi"
                      }
                    ],
                    "availability": null
                  },
                  "checkIn": "2026-07-01",
                  "checkOut": "2026-07-05",
                  "nights": 4,
                  "firstName": "Max",
                  "lastName": "Mustermann",
                  "email": "max@example.com",
                  "totalPrice": 626.0,
                  "status": "CONFIRMED",
                  "createdAt": "2026-05-04T14:30:00Z",
                  "breakfast": true
                }
              }
            }
          },
          "400": {
            "description": "Validierungsfehler",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "example": {
                  "timestamp": "2026-05-04T14:30:00Z",
                  "status": 400,
                  "error": "Bad Request",
                  "message": "Validation failed",
                  "fieldErrors": [
                    {
                      "field": "email",
                      "message": "must be a valid email address"
                    }
                  ],
                  "path": "/api/v1/bookings"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/bookings/{id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Buchungsbestätigung abrufen",
        "description": "Gibt die vollständige Buchungsbestätigung zurück inkl. Hotelinformationen.",
        "operationId": "getBookingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID der Buchung",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Buchung gefunden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfirmation"
                },
                "example": {
                  "id": "a3f1c2d4-e5b6-7890-abcd-ef1234567890",
                  "room": {
                    "id": 1,
                    "title": "Deluxe Suite",
                    "description": "Geräumige Suite mit Balkon und Bergblick.",
                    "imageUrl": "https://hotel-technikum.at/images/rooms/1.jpg",
                    "pricePerNight": 149.0,
                    "extras": [
                      {
                        "id": 1,
                        "name": "WiFi",
                        "icon": "wifi"
                      }
                    ],
                    "availability": null
                  },
                  "checkIn": "2026-07-01",
                  "checkOut": "2026-07-05",
                  "nights": 4,
                  "firstName": "Max",
                  "lastName": "Mustermann",
                  "email": "max@example.com",
                  "totalPrice": 626.0,
                  "status": "CONFIRMED",
                  "createdAt": "2026-05-04T14:30:00Z",
                  "hotel": {
                    "name": "Boutique Hotel Technikum",
                    "address": "Höchstädtplatz 6, 1200 Wien",
                    "phone": "+43 1 XXXXXXX",
                    "email": "info@hotel-technikum.at",
                    "directions": "U6 Station Dresdner Straße, 2 Minuten Fußweg",
                    "coordinates": {
                      "lat": 48.2392,
                      "lng": 16.378
                    }
                  },
                  "breakfast": true
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "tags": [
          "Bookings"
        ],
        "summary": "Buchung stornieren",
        "description": "Storniert eine bestehende Buchung. Eine bereits stornierte Buchung kann nicht erneut storniert werden (409).",
        "operationId": "cancelBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID der Buchung",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "CANCELLED"
                  }
                },
                "required": [
                  "status"
                ]
              },
              "example": {
                "status": "CANCELLED"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Buchung erfolgreich storniert",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfirmation"
                },
                "example": {
                  "id": "a3f1c2d4-e5b6-7890-abcd-ef1234567890",
                  "room": {
                    "id": 1,
                    "title": "Deluxe Suite",
                    "description": "Geräumige Suite mit Balkon und Bergblick.",
                    "imageUrl": "https://hotel-technikum.at/images/rooms/1.jpg",
                    "pricePerNight": 149.0,
                    "extras": [
                      {
                        "id": 1,
                        "name": "WiFi",
                        "icon": "wifi"
                      }
                    ],
                    "availability": null
                  },
                  "checkIn": "2026-07-01",
                  "checkOut": "2026-07-05",
                  "nights": 4,
                  "firstName": "Max",
                  "lastName": "Mustermann",
                  "email": "max@example.com",
                  "totalPrice": 626.0,
                  "status": "CANCELLED",
                  "createdAt": "2026-05-04T14:30:00Z",
                  "hotel": {
                    "name": "Boutique Hotel Technikum",
                    "address": "Höchstädtplatz 6, 1200 Wien",
                    "phone": "+43 1 XXXXXXX",
                    "email": "info@hotel-technikum.at",
                    "directions": "U6 Station Dresdner Straße, 2 Minuten Fußweg",
                    "coordinates": {
                      "lat": 48.2392,
                      "lng": 16.378
                    }
                  },
                  "breakfast": true
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Buchung ist bereits storniert",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "timestamp": "2026-05-04T14:30:00Z",
                  "status": 409,
                  "error": "Conflict",
                  "message": "Booking is already cancelled",
                  "path": "/api/v1/bookings/a3f1c2d4-e5b6-7890-abcd-ef1234567890"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Extra": {
        "type": "object",
        "description": "Zimmer-Ausstattungsmerkmal",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "WiFi"
          },
          "icon": {
            "type": "string",
            "example": "wifi"
          }
        },
        "required": [
          "id",
          "name",
          "icon"
        ]
      },
      "RoomAvailabilitySummary": {
        "type": "object",
        "description": "Eingebettete Verfügbarkeitsinfo – nur gesetzt wenn GET /rooms mit checkIn/checkOut aufgerufen wurde",
        "properties": {
          "available": {
            "type": "boolean",
            "example": true
          },
          "nights": {
            "type": "integer",
            "example": 4
          },
          "totalPrice": {
            "type": "number",
            "format": "double",
            "nullable": true,
            "description": "Gesamtpreis in EUR. Null wenn nicht verfügbar.",
            "example": 596.0
          }
        },
        "required": [
          "available",
          "nights"
        ]
      },
      "Room": {
        "type": "object",
        "description": "Hotelzimmer",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Deluxe Suite"
          },
          "description": {
            "type": "string",
            "example": "Geräumige Suite mit Balkon und Bergblick."
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "description": "Absolute URL zum Zimmerbild",
            "example": "https://hotel-technikum.at/images/rooms/1.jpg"
          },
          "pricePerNight": {
            "type": "number",
            "format": "double",
            "example": 149.0
          },
          "extras": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Extra"
            }
          },
          "availability": {
            "nullable": true,
            "description": "Nur befüllt wenn GET /rooms mit checkIn und checkOut aufgerufen wurde. Sonst null.",
            "allOf": [
              {
                "$ref": "#/components/schemas/RoomAvailabilitySummary"
              }
            ]
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "imageUrl",
          "pricePerNight",
          "extras"
        ]
      },
...
      "BookingRequest": {
        "type": "object",
        "description": "Anfrage zur Erstellung einer Buchung. Der totalPrice wird vom Backend berechnet und beinhaltet ggf. den Frühstücksaufpreis.",
        "properties": {
          "roomId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "checkIn": {
            "type": "string",
            "format": "date",
            "example": "2026-07-01"
          },
          "checkOut": {
            "type": "string",
            "format": "date",
            "description": "Abreisedatum. Darf gleich checkIn sein (= 1 Nacht). Maximaler Zeitraum: 30 Nächte.",
            "example": "2026-07-05"
          },
          "firstName": {
            "type": "string",
            "maxLength": 100,
            "example": "Max"
          },
          "lastName": {
            "type": "string",
            "maxLength": 100,
            "example": "Mustermann"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "max@example.com"
          },
          "breakfast": {
            "type": "boolean",
            "description": "Frühstück gewünscht. Beeinflusst den totalPrice – der genaue Aufpreis wird vom Backend berechnet.",
            "example": true
          }
        },
        "required": [
          "roomId",
          "checkIn",
          "checkOut",
          "firstName",
          "lastName",
          "email",
          "breakfast"
        ]
      },
      "BookingStatus": {
        "type": "string",
        "description": "Status der Buchung",
        "enum": [
          "CONFIRMED",
          "CANCELLED"
        ],
        "example": "CONFIRMED"
      },
      "BookingResponse": {
        "type": "object",
        "description": "Erstellte oder abgerufene Buchung",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Eindeutige UUID der Buchung",
            "example": "a3f1c2d4-e5b6-7890-abcd-ef1234567890"
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          },
          "checkIn": {
            "type": "string",
            "format": "date",
            "example": "2026-07-01"
          },
          "checkOut": {
            "type": "string",
            "format": "date",
            "example": "2026-07-05"
          },
          "nights": {
            "type": "integer",
            "example": 4
          },
          "firstName": {
            "type": "string",
            "example": "Max"
          },
          "lastName": {
            "type": "string",
            "example": "Mustermann"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "max@example.com"
          },
          "totalPrice": {
            "type": "number",
            "format": "double",
            "description": "Gesamtpreis in EUR inkl. Frühstücksaufpreis wenn gebucht.",
            "example": 596.0
          },
          "breakfast": {
            "type": "boolean",
            "description": "Gibt an ob Frühstück gebucht wurde.",
            "example": true
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Erstellungszeitpunkt der Buchung. Immer in UTC (Z).",
            "example": "2026-05-04T14:30:00Z"
          }
        },
        "required": [
          "id",
          "room",
          "checkIn",
          "checkOut",
          "nights",
          "firstName",
          "lastName",
          "email",
          "totalPrice",
          "breakfast",
          "status",
          "createdAt"
        ]
      },
      "HotelCoordinates": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "format": "double",
            "example": 48.2392
          },
          "lng": {
            "type": "number",
            "format": "double",
            "example": 16.378
          }
        },
        "required": [
          "lat",
          "lng"
        ]
      },
      "HotelInfo": {
        "type": "object",
        "description": "Hotelinformationen für die Buchungsbestätigung",
        "properties": {
          "name": {
            "type": "string",
            "example": "Boutique Hotel Technikum"
          },
          "address": {
            "type": "string",
            "example": "Höchstädtplatz 6, 1200 Wien"
          },
          "phone": {
            "type": "string",
            "example": "+43 1 XXXXXXX"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "info@hotel-technikum.at"
          },
          "directions": {
            "type": "string",
            "example": "U6 Station Dresdner Straße, 2 Minuten Fußweg"
          },
          "coordinates": {
            "$ref": "#/components/schemas/HotelCoordinates"
          }
        },
        "required": [
          "name",
          "address",
          "phone",
          "email",
          "directions",
          "coordinates"
        ]
      },
      "BookingConfirmation": {
        "description": "Vollständige Buchungsbestätigung inkl. Hotelinformationen",
        "allOf": [
          {
            "$ref": "#/components/schemas/BookingResponse"
          },
          {
            "type": "object",
            "properties": {
              "hotel": {
                "$ref": "#/components/schemas/HotelInfo"
              }
            },
            "required": [
              "hotel"
            ]
          }
        ]
      },
      "ApiError": {
        "type": "object",
        "description": "Einheitliches Fehlerformat",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Fehlerzeitpunkt in UTC",
            "example": "2026-05-04T14:30:00Z"
          },
          "status": {
            "type": "integer",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "checkOut must be after checkIn"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/rooms/1/availability"
          }
        },
        "required": [
          "timestamp",
          "status",
          "error",
          "message",
          "path"
        ]
      },
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "email"
          },
          "message": {
            "type": "string",
            "example": "must be a valid email address"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ValidationError": {
        "description": "Validierungsfehler mit Felddetails",
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiError"
          },
          {
            "type": "object",
            "properties": {
              "fieldErrors": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                }
              }
            }
          }
        ]
      }
    },
    "responses": {
...
      "NotFound": {
        "description": "Ressource nicht gefunden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-05-04T14:30:00Z",
              "status": 404,
              "error": "Not Found",
              "message": "Room not found",
              "path": "/api/v1/rooms/99"
            }
          }
        }
      },
      "Conflict": {
        "description": "Zimmer im gewünschten Zeitraum bereits gebucht",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-05-04T14:30:00Z",
              "status": 409,
              "error": "Conflict",
              "message": "Room is no longer available for the selected dates",
              "path": "/api/v1/bookings"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Interner Serverfehler",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-05-04T14:30:00Z",
              "status": 500,
              "error": "Internal Server Error",
              "message": "An unexpected error occurred",
              "path": "/api/v1/rooms"
            }
          }
        }
      }
    }
  }
}